refactoring pages and prop types
